Move FeatureFlags type to lib-common
