+++ frontend/src/lib-customizations/oulu/featureFlags.tsx
@@ -6,7 +6,7 @@ SPDX-License-Identifier: LGPL-2.1-or-later
 */
 }
 
-import type { FeatureFlags } from 'lib-customizations/types'
+import type { FeatureFlags } from 'lib-common/feature-flags'
 
 import { env } from './env'
 import type { Env } from './env'
